+ proper comment and fix type error

+++ src/Player.java
@@ -15,13 +15,14 @@ public abstract class Player {
 	public static final byte COMMONER = 3;
 	public static final byte NEEDY = 4;
 	public static final byte EXTREME_NEEDY = 5;
-	private int title;
+	private byte title;
 	
 
 	// ----- actions ----- //
 
 	/**
 	 * to increase the number of points this player has
+	 * @param the amount of points to add
 	 */
 	public final void win_points(int amount) {
 		point += amount;
@@ -29,26 +30,39 @@ public abstract class Player {
 	
 	/**
 	 * to decrease the number of points this player has
+	 * @param the amount of points to decrease
 	 */
 	public final void lose_points(int amount) {
 		point -= amount;
 	}
 
+	/**
+	 * @return the total points this player has
+	 */
 	public final int get_points() {
 		int ret = point;
 		return ret;
 	}
 
+	/**
+	 * to set the title of this player
+	 * @param t the title
+	 */
 	public final void set_title(byte t) {
 		title = t;
 	}
 
+	/**
+	 * to get the title of this player
+	 * @param the title of this player
+	 */
 	public final byte get_title() {
 		byte ret = title;
 		return ret;
 	}
 
 	// ----- abstract methods ----- //
+
 	public abstract ArrayList<Card> play_card(ArrayList<Card> myCards);
 	public abstract void print_message(Message msg);
 }
